+++ src/io.h
@@ -8,8 +8,8 @@ inline void PrintMove(const Move move) {
     const char* to = square_to_coordinates[To(move)];
 
     if (isPromo(move))
-        printf("%s%s%c", from, to, promoted_pieces(getPromotedPiecetype(move)));
+        printf("%s%s%c", (const size_t)from, (const size_t)to, promoted_pieces(getPromotedPiecetype(move)));
     else
-        printf("%s%s", from, to);
+        printf("%s%s", (const size_t)from, (const size_t)to);
 }
